Don't strictly fail on dossier update if its not found
Sometimes given PM or PP for updates do not exist in the stock... So we
now create them instead. See issue #36 for more details.
Therefore numero_gestion (which is part of the key to identify the
dossier to update) can change (in case of forme_juridique change for
example) so we now have a way to update it.