Documenting trailblazer edge comportment
The :success track ends at :update step. The :create task is then
floating in an outer dimension, only accessible from another task
pointing to it with the :id option and therefore not being able to end
to the usual :success or :failure end events. Without specifying the
Output for the task you get stuck in an infinite loop where the :create
task runs forever.