Don't fail import when encountering an unknown PM or PP
During import, représentants, établissements and observations may refer
to a PM or PP (a "dossier" in our model) which is unknown (which has
never been transmitted and therefor is not present in the DB).
In such a case, we ignore the line and log the record could not be
saved then continue the import.

Related to issue #39